requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.10.3
mercantile==1.2.1
vt2geojson==0.2.1
shapely==2.0.6
numpy>=1.26
opencv-python>=4.10
py360convert>=1.0.3
Pillow>=10.4
anthropic>=0.39
